Fix hash and clean up filter code
