+++ public/index.html
@@ -1,16 +1,17 @@
 <!DOCTYPE html>
 <html>
-  <head>
-    <meta charset="utf-8" />
-    <link rel="shortcut icon" href="%PUBLIC_URL%/favicon.ico" />
-    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no" />
-    <meta name="theme-color" content="#000000" />
-    <!--
+
+<head>
+  <meta charset="utf-8" />
+  <link rel="shortcut icon" href="%PUBLIC_URL%/favicon.ico" />
+  <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no" />
+  <meta name="theme-color" content="#000000" />
+  <!--
       manifest.json provides metadata used when your web app is installed on a
       user's mobile device or desktop. See https://developers.google.com/web/fundamentals/web-app-manifest/
     -->
-    <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
-    <!--
+  <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
+  <!--
       Notice the use of %PUBLIC_URL% in the tags above.
       It will be replaced with the URL of the `public` folder during the build.
       Only files inside the `public` folder can be referenced from the HTML.
@@ -19,22 +20,23 @@
       work correctly both with client-side routing and a non-root public URL.
       Learn how to configure a non-root public URL by running `npm run build`.
     -->
-    <title>Trelliscope</title>
-  </head>
+  <title>Trelliscope</title>
+</head>
 
-  <body>
-    <noscript>You need to enable JavaScript to run this app.</noscript>
-    <!-- <div id="96c61ca5" class="trelliscope-not-spa" style="width: 800px; height: 500px"></div> -->
-    <div id="44c922eb" class="trelliscope-spa"></div>
-    <script>
-      (function () {
-        // trelliscopeApp('6c048a7', '_test/example_gapminder_plotly/config.jsonp', { logger: true });
-        // trelliscopeApp('32b2482a', '_test/example_gapminder/config.jsonp', { logger: true });
-        // trelliscopeApp('d27693de', '_test/pc_ratio/config.jsonp', { logger: false });
-        // trelliscopeApp('8a43f2dd', '_test/example_housing/config.jsonp');
-      })();
-    </script>
-  </body>
+<body>
+  <noscript>You need to enable JavaScript to run this app.</noscript>
+  <!-- <div id="96c61ca5" class="trelliscope-not-spa" style="width: 800px; height: 500px"></div> -->
+  <div id="960f51e6" class="trelliscope-spa"></div>
+  <script>
+    (function () {
+      // trelliscopeApp('6c048a7', '_test/example_gapminder_plotly/config.jsonp', { logger: true });
+      // trelliscopeApp('32b2482a', '_test/example_gapminder/config.jsonp', { logger: true });
+      // trelliscopeApp('d27693de', '_test/pc_ratio/config.jsonp', { logger: false });
+      // trelliscopeApp('8a43f2dd', '_test/example_housing/config.jsonp');
+    })();
+  </script>
+</body>
+
 </html>
 
 <!-- <!doctype html>
