+++ src/components/SidebarFilter/FilterInput/FilterCat/FilterCat.module.scss
