+++ src/components/SidebarFilter/FilterInput/FilterInput.module.scss
