+++ src/components/SidebarFilter/SidebarFilter.module.scss
@@ -1,28 +1,1 @@
-.sidebarFilter {
-  &Col1 {
-    position: absolute;
-    top: 25px;
-    left: 0;
-  }
-
-  &Col2 {
-    position: absolute;
-    top: 25px;
-    left: 230px;
-    border-left: var(--border-6);
-  }
-
-  &NotUsedHeader {
-    height: var(--sidebarFilter-notUsedHeader-height);
-    line-height: var(--sidebarFilter-notUsedHeader-line-height);
-    padding-left: var(--padding-7);
-    box-sizing: border-box;
-    width: var(--sidebarFilter-width);
-    font-size: var(--text-base);
-  }
-
-  &AllContainer {
-    width: var(--sidebarFilter-width);
-    box-sizing: border-box;
-  }
-}
+.sidebarFilter {}
