--- src/components/SidebarFilterNew/FilterList/FilterList.tsx
@@ -1,43 +1,0 @@
-import React from 'react';
-import { useDispatch } from 'react-redux';
-import { commonTagsKey, setFilter, setFilterView, useInactiveFilterGroups } from '../../../slices/filterSlice';
-import Pill from '../../Pill';
-
-import styles from './FilterList.module.scss';
-
-interface FilterListProps {}
-
-const FilterList: React.FC<FilterListProps> = () => {
-  const inactiveFilterGroups = useInactiveFilterGroups();
-  const dispatch = useDispatch();
-
-  const handleClick = (filter: string) => {
-    dispatch(setFilterView({ name: filter, which: 'add' }));
-  };
-
-  return (
-    <div className={styles.filterList}>
-      <div className={styles.filterListHeading}>Select a variable to filter on:</div>
-      {Object.keys(inactiveFilterGroups).map((key) => {
-        return (
-          <div className={styles.filterListGroup} key={key}>
-            {key !== commonTagsKey && (
-              <div className={styles.filterListGroupHeader}>
-                {key} ({inactiveFilterGroups[key].length})
-              </div>
-            )}
-            <ul className={styles.filterListItems}>
-              {inactiveFilterGroups[key].map((filter) => (
-                <Pill key={filter} onClick={() => handleClick(filter)}>
-                  {filter}
-                </Pill>
-              ))}
-            </ul>
-          </div>
-        );
-      })}
-    </div>
-  );
-};
-
-export default FilterList;
