--- src/components/SidebarFilterNew/SidebarFilterNew.module.scss
@@ -1,1 +1,0 @@
-.sidebarFilterNew {}
