+++ src/slices/displayInfoAPI.ts
@@ -1,3 +1,4 @@
+import { useMemo } from 'react';
 import { BaseQueryFn, createApi } from '@reduxjs/toolkit/query/react';
 import getJSONP from 'browser-jsonp';
 import { useSelector } from 'react-redux';
@@ -56,8 +57,8 @@ export const useDisplayInfo = () => {
 };
 
 export const useDisplayMetas = () => {
-  const displayInfo = useDisplayInfo();
-  return displayInfo.data?.metas || [];
+  const { data } = useDisplayInfo();
+  return useMemo(() => data?.metas || [], [data]);
 };
 
 export const useMetaByVarname = (varname: string) => {
